requirements.txt
pillow==9.5.0
numpy==1.24.3
pandas==2.0.2
torch==2.0.1
torchvision==0.15.2
google-colab==1.0.0